+++ mojo_opset/backends/ttx/operators/mrope.py
@@ -5,7 +5,7 @@ from typing import Tuple
 import torch
 
 from mojo_opset.backends.ttx.kernels.npu.mrope import mrope_fwd_impl
-from mojo_opset.core.operators.mrope import MojoMRoPE
+from mojo_opset.core.operators.position_embedding import MojoMRoPE
 
 
 class TTXMRoPE(MojoMRoPE):
