+++ mojo_opset/core/operators/position_embedding.py
@@ -7,11 +7,17 @@ from ..operator import MojoOperator
 
 
 class MojoRotaryEmbedding(MojoOperator):
-    def __init__(self, rope_theta, rope_dim, attention_scaling: float = 1.0, init_max_length: Optional[int] = None, **kwargs):
+    def __init__(
+        self, rope_theta, rope_dim, attention_scaling: float = 1.0, init_max_length: Optional[int] = None, **kwargs
+    ):
         super().__init__(**kwargs)
         self.rope_theta = rope_theta
         inv_freq = 1.0 / (
-            self.rope_theta ** (torch.arange(0, rope_dim, 2, dtype=torch.float32, device = self.tensor_factory_kwargs.get("device")) / rope_dim)
+            self.rope_theta
+            ** (
+                torch.arange(0, rope_dim, 2, dtype=torch.float32, device=self.tensor_factory_kwargs.get("device"))
+                / rope_dim
+            )
         )
         self.attention_scaling = attention_scaling
         self.init_max_length = None
@@ -22,16 +28,16 @@ class MojoRotaryEmbedding(MojoOperator):
         def load_state_dict_post_hook(module, incompatible_keys) -> None:
             key2ignore = []
             for miss in incompatible_keys.missing_keys:
-                if miss.split('.')[-1] in ("inv_freq", "cos", "sin"):
+                if miss.split(".")[-1] in ("inv_freq", "cos", "sin"):
                     key2ignore.append(miss)
             for key in key2ignore:
                 incompatible_keys.missing_keys.remove(key)
-        self.register_load_state_dict_post_hook(load_state_dict_post_hook)
 
+        self.register_load_state_dict_post_hook(load_state_dict_post_hook)
 
     def _rope_init(self, max_length: Optional[int]) -> Tuple[torch.Tensor, torch.Tensor]:
         self.init_max_length = max_length
-        position_ids = torch.arange(max_length, device = self.tensor_factory_kwargs.get("device"))
+        position_ids = torch.arange(max_length, device=self.tensor_factory_kwargs.get("device"))
         freqs = position_ids[..., None] * self.inv_freq[None, :]
         emb = torch.cat((freqs, freqs), dim=-1)
         cos = emb.cos() * self.attention_scaling
@@ -55,26 +61,30 @@ class MojoRotaryEmbedding(MojoOperator):
         2. Padded prefill: input [B, S, H], cu_seqlens_q None, position_ids None -> cos/sin [S, D].
         3. Decode: input [B, H], cu_seqlens_q None, position_ids [B] -> cos/sin [B, D].
         """
-        assert position_ids is None or cu_seqlens_q is None, "At most one of cu_seqlens_q or position_ids should be provided"
+        assert position_ids is None or cu_seqlens_q is None, (
+            "At most one of cu_seqlens_q or position_ids should be provided"
+        )
 
         if cu_seqlens_q is not None:
             assert x.dim() == 2, "x must be 2D: [T, D]"
-            position_ids = torch.full((x.shape[0],), -1, device = x.device, dtype = torch.int32)
+            position_ids = torch.full((x.shape[0],), -1, device=x.device, dtype=torch.int32)
             seqlens_q = cu_seqlens_q[1:] - cu_seqlens_q[:-1]
             bsz = seqlens_q.size(0)
             for i in range(bsz):
                 q_len = seqlens_q[i].item()
                 context_len = 0 if seqlens_kv is None else seqlens_kv[i].item() - q_len
-                position_ids[cu_seqlens_q[i]:cu_seqlens_q[i+1]] = torch.arange(
+                position_ids[cu_seqlens_q[i] : cu_seqlens_q[i + 1]] = torch.arange(
                     context_len,
-                    context_len + q_len, 
-                    device = cu_seqlens_q.device,
-                    dtype = torch.int32,
+                    context_len + q_len,
+                    device=cu_seqlens_q.device,
+                    dtype=torch.int32,
                 )
         elif position_ids is not None:
-            assert position_ids.shape == x.shape[:-1], "position_ids must have the same shape as x except the hidden dimension"
+            assert position_ids.shape == x.shape[:-1], (
+                "position_ids must have the same shape as x except the hidden dimension"
+            )
         else:
-            position_ids = torch.arange(x.shape[1], device = x.device, dtype = torch.int32)
+            position_ids = torch.arange(x.shape[1], device=x.device, dtype=torch.int32)
 
         if self.init_max_length is None:
             freqs = position_ids[..., None] * self.inv_freq[None, :]
@@ -84,12 +94,11 @@ class MojoRotaryEmbedding(MojoOperator):
         else:
             cos = self.cos[position_ids]
             sin = self.sin[position_ids]
-        
+
         return cos, sin
 
 
 class MojoApplyRoPE(MojoOperator):
-
     def __init__(self, interleaved: bool = False):
         super().__init__()
         assert not interleaved, "interleaved impl is not supported yet."
@@ -157,8 +166,10 @@ class MojoApplyRoPE(MojoOperator):
         assert q.ndim == 3 or q.ndim == 4, "q and k must be 3D or 4D"
         assert cos.shape == sin.shape, "cos and sin must have the same shape"
         if q.ndim == 3:
-            assert cos.ndim == 2, "rotary position embedding (cos/sin) must be of shape [num_tokens, rope_dim] for varlen prefill or decode"
-        
+            assert cos.ndim == 2, (
+                "rotary position embedding (cos/sin) must be of shape [num_tokens, rope_dim] for varlen prefill or decode"
+            )
+
         if head_first:
             cos = cos.unsqueeze(-3)
             sin = sin.unsqueeze(-3)
@@ -219,3 +230,126 @@ class MojoGridRoPE(MojoOperator):
             output.append(x_i)
         y = torch.stack(output)
         return y.type_as(x)
+
+
+class MojoMRoPE(MojoOperator):
+    """Multimodal Rotary Position Embedding (MRoPE) for Qwen2-VL.
+
+    Applies 3D rotary position embedding over temporal (T), height (H), and width (W)
+    dimensions to query and key tensors. Supports both interleaved and non-interleaved modes.
+
+    Reference: https://qwenlm.github.io/blog/qwen2-vl/
+    """
+
+    supported_platforms_list = ["npu"]
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+    def extra_repr(self) -> str:
+        return ""
+
+    @staticmethod
+    def _rotate_half(hidden_states: torch.Tensor) -> torch.Tensor:
+        """Rotates half the hidden dims of the input."""
+        hidden_size = hidden_states.shape[-1]
+        hidden_states_half = hidden_size // 2
+        left = hidden_states[..., :hidden_states_half]
+        right = hidden_states[..., hidden_states_half:]
+        return torch.cat((-right, left), dim=-1)
+
+    @staticmethod
+    def _apply_interleaved_mrope(
+        cos_table: torch.Tensor,
+        sin_table: torch.Tensor,
+        mrope_section: List[int],
+    ) -> Tuple[torch.Tensor, torch.Tensor]:
+        """Apply interleaved MRoPE pattern to cos/sin tables."""
+        cos_interleaved = cos_table[0].clone()
+        cos_interleaved[..., 1::3] = cos_table[1, ..., 1::3]
+        cos_interleaved[..., 2::3] = cos_table[2, ..., 2::3]
+
+        sin_interleaved = sin_table[0].clone()
+        sin_interleaved[..., 1::3] = sin_table[1, ..., 1::3]
+        sin_interleaved[..., 2::3] = sin_table[2, ..., 2::3]
+
+        return cos_interleaved, sin_interleaved
+
+    def forward(
+        self,
+        query: torch.Tensor,
+        key: torch.Tensor,
+        cos_table: torch.Tensor,
+        sin_table: torch.Tensor,
+        mrope_section: List[int],
+        is_interleaved: bool = False,
+        head_dim: Optional[int] = None,
+    ) -> Tuple[torch.Tensor, torch.Tensor]:
+        """
+        Apply Multimodal Rotary Position Embedding to query and key tensors.
+
+        Args:
+            query: ``(num_tokens, n_qh * head_dim)`` query tensor.
+            key: ``(num_tokens, n_kh * head_dim)`` key tensor.
+            cos_table: ``(3, num_tokens, rotary_dim // 2)`` cos values for T/H/W dimensions.
+            sin_table: ``(3, num_tokens, rotary_dim // 2)`` sin values for T/H/W dimensions.
+            mrope_section: ``[t_section, h_section, w_section]`` - how half rope_dim is split.
+            is_interleaved: if True, T/H/W positions are interleaved.
+            head_dim: head dimension. If None, inferred from cos_table.
+
+        Returns:
+            ``(query, key)`` with RoPE applied, same shape as input.
+        """
+        num_tokens, n_qh_head_dim = query.shape
+        num_tokens_k, n_kh_head_dim = key.shape
+
+        rope_dim = sum(mrope_section) * 2
+        half_rope_dim = rope_dim // 2
+
+        if head_dim is None:
+            head_dim = cos_table.shape[-1] * 2
+            if head_dim != rope_dim:
+                raise ValueError(
+                    f"head_dim ({head_dim}) inferred from cos_table does not match "
+                    f"rope_dim ({rope_dim}). Please pass head_dim explicitly."
+                )
+
+        n_qh = n_qh_head_dim // head_dim
+        n_kh = n_kh_head_dim // head_dim
+
+        query = query.view(num_tokens, n_qh, head_dim)
+        key = key.view(num_tokens_k, n_kh, head_dim)
+
+        query_rot, query_pass = query.split([rope_dim, head_dim - rope_dim], dim=-1)
+        key_rot, key_pass = key.split([rope_dim, head_dim - rope_dim], dim=-1)
+
+        if cos_table.dim() == 3:
+            if is_interleaved:
+                cos_table, sin_table = self._apply_interleaved_mrope(cos_table, sin_table, mrope_section)
+            else:
+                cos_table = torch.cat([m[i] for i, m in enumerate(cos_table.split(mrope_section, dim=-1))], dim=-1)
+                sin_table = torch.cat([m[i] for i, m in enumerate(sin_table.split(mrope_section, dim=-1))], dim=-1)
+
+        cos_table = cos_table.view(num_tokens, half_rope_dim)
+        sin_table = sin_table.view(num_tokens, half_rope_dim)
+
+        query_rot_half1 = query_rot[..., :half_rope_dim]
+        query_rot_half2 = query_rot[..., half_rope_dim:]
+        key_rot_half1 = key_rot[..., :half_rope_dim]
+        key_rot_half2 = key_rot[..., half_rope_dim:]
+
+        cos_expanded = cos_table.unsqueeze(1)
+        sin_expanded = sin_table.unsqueeze(1)
+
+        query_rot_new_half1 = query_rot_half1 * cos_expanded - query_rot_half2 * sin_expanded
+        query_rot_new_half2 = query_rot_half2 * cos_expanded + query_rot_half1 * sin_expanded
+        key_rot_new_half1 = key_rot_half1 * cos_expanded - key_rot_half2 * sin_expanded
+        key_rot_new_half2 = key_rot_half2 * cos_expanded + key_rot_half1 * sin_expanded
+
+        query_rot = torch.cat([query_rot_new_half1, query_rot_new_half2], dim=-1)
+        key_rot = torch.cat([key_rot_new_half1, key_rot_new_half2], dim=-1)
+
+        query = torch.cat([query_rot, query_pass], dim=-1).view(num_tokens, -1)
+        key = torch.cat([key_rot, key_pass], dim=-1).view(num_tokens_k, -1)
+
+        return query, key
